Move one-command reproduction out to the research repo
wmo/reproduce (manifests, runner, CLI) now lives in experientiallabs/research,
which imports this package for all plumbing; wmo keeps no reproduce command.
CachedTaskEmbedder stays here — it is plumbing, not reproduction: the DeepSWE
converter and its tests consume it — relocated to wmo/optimize/embedding_cache.py.
Docs now name the research repo's reproduce command wherever they quoted
wmo reproduce.
