+++ docs/README.md
@@ -21,11 +21,11 @@ here must justify its existence in the table below; a doc that can't gets delete
 |---|---|
 | `README.md` | The manifest that makes the justification rule enforceable. |
 | `usage.md` | The one-page map of the CLI surface: every command's purpose and the artifact it leaves behind, grouped by pipeline / optimizers / traces / platform. The reference-style counterpart to the cookbook's narrative order, and the page that keeps every other doc from re-explaining what a command is. |
-| `cookbook/tau-bench.md` | The canonical end-to-end example: one pass through the whole pipeline (setup, build, pool, optimize, optional distill, serve) told on tau-bench, each step one real CLI command plus the artifact it creates. Commands verified against `main`; the provenance rules (world-model simulated vs real-episode, cache-adjusted effective cost per completed task, savings as estimates) are stated once here and inherited by every number in the walk. Ends with the measured 2026-07-28 results (routed vs anchor at both dial points, traffic mix, caveats), the one-command protocol-exact reproduction (`wmo reproduce run tau-bench --yes`), and the real-episode leg: the shipped tau default re-measured on Sierra's actual benchmark, with its own free bit-exact reproduction (`wmo reproduce run tau-bench-real`) and a provenance rule that the two legs' numbers never blend. |
-| `cookbook/routerbench.md` | The shortest routing walk: fit + tune + report on a precomputed RouterBench-style outcome matrix, no world model in the loop. Documents the matrix shape, the guarded kNN fit, the held-out report, the measured `ours9` results (both fits, both anchors, concrete routed requests with the guard's verbatim reasoning), and the one-command bit-exact reproduction (`wmo reproduce run routerbench`) against the published dataset. |
-| `cookbook/terminal-tasks.md` | Terminal work in two legs, interpreted together and never blended. The public end-to-end walk on the `wmh-terminal-tasks-traces` Hugging Face bundle (measured constant-policy result, why the dial saturates, the compression cost-inversion), then the real Terminal-Bench 2.0 grid behind the product's terminal default: 640 real episodes, the pinned sonnet-5 table, the sim-to-real caveats (aggregate agreement, per-scenario rho -0.162), and the free bit-exact reproduction (`wmo reproduce run terminal-bench-2`). |
-| `cookbook/swe-bench.md` | The evidence page behind the product's `swe-bench` default: 640 real SWE-bench Verified episodes, the pinned opus-5 table with its resolved-cost/unresolved-quality reading, the router null result and its mechanism, the cautionary rows (cheapest tokens = dearest completions; per-arm cap concentration), and the free bit-exact reproduction (`wmo reproduce run swe-bench`) from the published matrix. Deliberately a replay page, not a pipeline walk: the grid harness that bought the episodes is not a product CLI step, and the page says so. |
-| `cookbook/deepswe.md` | The credential-free routing walk: DeepSWE v1.1's published long-horizon SWE trials converted into an outcome matrix (`wmo optimize route convert-deepswe`, gated on reproducing every published pass@1), the in-process local embedder (Qwen3-Embedding-0.6B via MLX/torch, no embedding API), the repo-grouped holdout protocol with its measured dial frontier beside the pre-split lab's threshold-rule reference, and the free bit-exact reproduction (`wmo reproduce run deepswe-coding113`). |
+| `cookbook/tau-bench.md` | The canonical end-to-end example: one pass through the whole pipeline (setup, build, pool, optimize, optional distill, serve) told on tau-bench, each step one real CLI command plus the artifact it creates. Commands verified against `main`; the provenance rules (world-model simulated vs real-episode, cache-adjusted effective cost per completed task, savings as estimates) are stated once here and inherited by every number in the walk. Ends with the measured 2026-07-28 results (routed vs anchor at both dial points, traffic mix, caveats), the one-command protocol-exact reproduction (`reproduce run tau-bench --yes` in the research repo), and the real-episode leg: the shipped tau default re-measured on Sierra's actual benchmark, with its own free bit-exact reproduction (`reproduce run tau-bench-real` in the research repo) and a provenance rule that the two legs' numbers never blend. |
+| `cookbook/routerbench.md` | The shortest routing walk: fit + tune + report on a precomputed RouterBench-style outcome matrix, no world model in the loop. Documents the matrix shape, the guarded kNN fit, the held-out report, the measured `ours9` results (both fits, both anchors, concrete routed requests with the guard's verbatim reasoning), and the one-command bit-exact reproduction (`reproduce run routerbench` in the research repo) against the published dataset. |
+| `cookbook/terminal-tasks.md` | Terminal work in two legs, interpreted together and never blended. The public end-to-end walk on the `wmh-terminal-tasks-traces` Hugging Face bundle (measured constant-policy result, why the dial saturates, the compression cost-inversion), then the real Terminal-Bench 2.0 grid behind the product's terminal default: 640 real episodes, the pinned sonnet-5 table, the sim-to-real caveats (aggregate agreement, per-scenario rho -0.162), and the free bit-exact reproduction (`reproduce run terminal-bench-2` in the research repo). |
+| `cookbook/swe-bench.md` | The evidence page behind the product's `swe-bench` default: 640 real SWE-bench Verified episodes, the pinned opus-5 table with its resolved-cost/unresolved-quality reading, the router null result and its mechanism, the cautionary rows (cheapest tokens = dearest completions; per-arm cap concentration), and the free bit-exact reproduction (`reproduce run swe-bench` in the research repo) from the published matrix. Deliberately a replay page, not a pipeline walk: the grid harness that bought the episodes is not a product CLI step, and the page says so. |
+| `cookbook/deepswe.md` | The credential-free routing walk: DeepSWE v1.1's published long-horizon SWE trials converted into an outcome matrix (`wmo optimize route convert-deepswe`, gated on reproducing every published pass@1), the in-process local embedder (Qwen3-Embedding-0.6B via MLX/torch, no embedding API), the repo-grouped holdout protocol with its measured dial frontier beside the pre-split lab's threshold-rule reference, and the free bit-exact reproduction (`reproduce run deepswe-coding113` in the research repo). |
 | `cookbook/fully_local.md` | The zero-cloud walk: two Ollama models as routing candidates, the simulation and judge on a local model via OPENAI_BASE_URL, sweep + knn fit + `wmo serve`, all at $0; includes where traces come from (your stack or `wmo download <benchmark>`). |
 | `research/world_model_findings.md` | The single research record: six layered studies (data, retrieval, optimization, test-time compute, self-knowledge, economics; PRs #72, #97, #55, #120, #41, with #83/#98 as instruments) with shared protocol and judge provenance stated once. Every product claim about world-model fidelity and cost traces to a section of this document. |
 | `research/figures/trace_scaling_law.png` | The trace-scaling figure (fidelity vs trace count, n=0 anchored) the record's data layer renders; also the brand-system visual reference cited by AGENTS.md rule 15. |
