+++ docs/usage.md
@@ -69,7 +69,7 @@ JSONL. The two formats are not interchangeable.
 |---|---|---|
 | `wmo login` / `logout` / `status` | Connect this machine to a platform account, disconnect, or show the current account and organizations. | a saved credential |
 | `wmo push` / `pull` | Publish a local world model to the platform registry, or fetch a model or endpoint state from it. | a registry entry, or a local artifact dir |
-| `wmo reproduce list` / `run <benchmark>` | Reproduce a published benchmark result from its shipped manifest: download the pinned data, replay the pinned protocol, and compare every published number field by field. `matrix` manifests run offline and bit-exact; `commands` manifests replay live CLI steps, state their estimated spend, and refuse without `--yes`. Exit 0 is REPRODUCED, 4 is DIVERGED. | `verdict.json` plus the run's own artifacts |
+| `reproduce list` / `run <benchmark>` (moved to the [research repo](https://github.com/experientiallabs/research)) | Reproduce a published benchmark result from its shipped manifest: download the pinned data, replay the pinned protocol, and compare every published number field by field. `matrix` manifests run offline and bit-exact; `commands` manifests replay live CLI steps, state their estimated spend, and refuse without `--yes`. Exit 0 is REPRODUCED, 4 is DIVERGED. | `verdict.json` plus the run's own artifacts |
 | `wmo runs list` / `show` / `tail` | See the runs this organization is feeding: progress, spend, stages, per-candidate cells, and the live event log. `--json` on the first two; `--org` (id or slug) to read another organization the credential can see, which otherwise comes from the login or `WMO_PLATFORM_ORG`. | nothing (prints) |
 | `wmo runs stop` / `retry` | Ask the process feeding a run to stop at its next safe boundary, or to re-measure its unscored cells. Pull-based: it takes effect when that process next reports in, and a runner that owns its own retry policy may refuse with a reason. | a queued command |
 | `wmo runs backfill <path>` | Replay a finished or interrupted run from its own artifacts (a grid directory, or a world model's `optimize/`), so a run nobody watched still has its history. The run's name comes from where the artifacts live; `--name` supplies it for artifacts that have moved. `--dry-run` writes the events as JSONL instead. | run history on the platform |
