--- wmo/reproduce/manifests/__init__.py
@@ -1,1 +1,0 @@
-"""Shipped reproduction manifests, one TOML per benchmark (see wmo.reproduce.manifest)."""
