--- wmo/reproduce/manifests/swe-bench.toml
@@ -1,31 +1,0 @@
-name = "swe-bench"
-title = "SWE-bench Verified: the pinned opus-5 default, measured on 640 real episodes"
-cookbook = "docs/cookbook/swe-bench.md"
-exactness = "bit-exact"
-kind = "matrix"
-
-notes = [
-  "640 real mini-swe-agent episodes: 20 pinned SWE-bench Verified instances x 16 candidate models x 2 episodes, graded by the benchmark's own test suite (no LLM judge).",
-  "Offline and credential-free: the shipped default is a static pin, which routes nothing, so the replay is arithmetic over the published matrix - no embeddings, no provider calls.",
-  "The 75-step cap (vs mini-swe-agent's shipped 250) applies uniformly; these solve rates are NOT comparable to published SWE-bench leaderboard numbers.",
-  "The quality delta vs the anchor (-2.5pt) is UNRESOLVED at n=20 (CI -10.0..+5.0, sign test p=1.00); the resolved claim is the cost saving.",
-]
-
-[data]
-hf_repo = "experiential-labs/wmo-swe-bench-defaults"
-repo_type = "dataset"
-revision = "112a3b9a8c44f71806b258f51f9eef7ef69a434a"
-files = ["matrix.json"]
-
-[matrix]
-matrix_file = "matrix.json"
-scenario_label = "on the 20 pinned SWE-bench Verified instances (real benchmark episodes, 2 per instance; 75-step cap; quality delta unresolved at this sample size)"
-pin_model = "opus-5"
-baselines = ["fable-5"]
-
-[[published]]
-label = "pinned opus-5 (the shipped swe default) vs anchor (fable-5)"
-baseline = "fable-5"
-accuracy = 0.85
-cost_per_run_usd = 0.7770521312499999
-latency_p50_ms = 6173.7557080050465
