--- wmo/reproduce/manifests/tau-bench.pool.toml
@@ -1,88 +1,0 @@
-# Reproduction pool for the tau-bench manifest: the measured 11-candidate roster on PUBLIC
-# routes. The measured run served five of these through private Azure deployments; here
-# they ride OpenRouter (self-priced from its live catalog), which keeps every entry callable
-# by a stranger with their own keys. Same models, sibling routes: reproduction on this pool
-# is protocol-exact, and price drift on the public routes moves cost columns accordingly.
-#
-# Keys: ANTHROPIC_API_KEY (five entries), OPENROUTER_API_KEY (five), FIREWORKS_API_KEY (one).
-
-[[model]]
-name = "gpt-5.5"
-kind = "openrouter"
-model = "openai/gpt-5.5"
-tier = "frontier"
-
-[[model]]
-name = "gpt-5.4-mini"
-kind = "openrouter"
-model = "openai/gpt-5.4-mini"
-tier = "frontier"
-
-[[model]]
-name = "deepseek-v4-pro"
-kind = "openrouter"
-model = "deepseek/deepseek-v4-pro"
-tier = "open"
-
-[[model]]
-name = "kimi-k2.6"
-kind = "openrouter"
-model = "moonshotai/kimi-k2.6"
-tier = "open"
-
-[[model]]
-name = "glm-5.2"
-kind = "openrouter"
-model = "z-ai/glm-5.2"
-tier = "open"
-
-[[model]]
-name = "fable-5"
-kind = "anthropic"
-model = "claude-fable-5"
-tier = "frontier"
-cached_input_per_mtok = 1.00
-
-[[model]]
-name = "sonnet-5"
-kind = "anthropic"
-model = "claude-sonnet-5"
-tier = "frontier"
-cached_input_per_mtok = 0.3
-
-[[model]]
-name = "haiku-4-5"
-kind = "anthropic"
-model = "claude-haiku-4-5"
-tier = "frontier"
-cached_input_per_mtok = 0.1
-
-[[model]]
-name = "opus-4-8"
-kind = "anthropic"
-model = "claude-opus-4-8"
-tier = "frontier"
-cached_input_per_mtok = 0.5
-
-[[model]]
-name = "opus-5"
-kind = "anthropic"
-model = "claude-opus-5"
-model_type = "claude-opus-5"
-tier = "frontier"
-input_per_mtok = 5.0
-output_per_mtok = 25.0
-cached_input_per_mtok = 0.5
-cache_write_per_mtok = 6.25
-
-[[model]]
-name = "kimi-k3"
-kind = "openai"
-model = "accounts/fireworks/models/kimi-k3"
-endpoint = "https://api.fireworks.ai/inference/v1"
-api_key_env = "FIREWORKS_API_KEY"
-tier = "open"
-input_per_mtok = 3.0
-output_per_mtok = 15.0
-cached_input_per_mtok = 0.3
-cache_write_per_mtok = 3.0
