--- wmo/reproduce/manifests/tau-bench.toml
@@ -1,40 +1,0 @@
-name = "tau-bench"
-title = "tau-bench: traces to routed endpoint, world-model closed loop"
-cookbook = "docs/cookbook/tau-bench.md"
-exactness = "protocol-exact"
-kind = "commands"
-
-notes = [
-  "Live-provider replay: world-model build plus a 440-cell sweep against real providers. Nondeterministic by nature; the tolerances state how close counts as reproduced.",
-  "The measured run served five candidates through private Azure deployments; this pool rides public routes for the same models, and public prices drift.",
-  "The world model builds on the anthropic provider here (the measured build used Bedrock; same model family, sibling route) so the whole replay runs on the pool's own documented keys.",
-  "Quality/cost are world-model simulated with the judge recorded on the artifacts; see docs/cookbook/tau-bench.md for provenance rules and caveats.",
-  "Estimated spend ~$110 at current prices (build + sweep + judge); the command refuses without --yes.",
-]
-
-[data]
-hf_repo = "experiential-labs/wmh-tau-bench-traces"
-revision = "5f1c44a2e0e80dec023665a3e8b10806be0ec0d2"
-repo_type = "dataset"
-files = ["traces.otel.jsonl"]
-
-[commands]
-pool_file = "tau-bench.pool.toml"
-estimated_spend_usd = 110.0
-report_file = "report.json"
-steps = [
-  ["build", "--file", "{data}/traces.otel.jsonl", "--name", "tau-bench-repro", "--provider", "anthropic", "--yes"],
-  ["optimize", "route", "sweep", "tau-bench-repro", "--pool", "{out}/pool.toml", "--scenarios", "20", "--episodes", "2", "--concurrency", "6", "--out", "{out}/matrix.json", "--yes"],
-  ["optimize", "route", "fit", "{out}/matrix.json", "--kind", "knn", "--out", "{out}/policy.json"],
-  ["optimize", "route", "report", "{out}/matrix.json", "{out}/policy.json", "--baseline", "fable-5", "--endpoint", "tau-bench-repro", "--out", "{out}/report.json"],
-]
-
-[[published]]
-label = "routed (balanced) vs anchor (fable-5), wm_simulated"
-accuracy = 0.5208333333333334
-cost_per_run_usd = 0.12588708333333334
-latency_p50_ms = 3159.7724159946665
-tolerance_accuracy = 0.20
-tolerance_cost = 0.35
-tolerance_latency = 0.50
-
